refactor: rename data mocks
