--- src/data/test/mock-http-post.ts
@@ -1,7 +1,0 @@
-import { faker } from '@faker-js/faker';
-import { HttpPostParams } from '../protocols/http';
-
-export const mockPostRequest = (): HttpPostParams<any> => ({
-  url: faker.internet.url(),
-  body: faker.getMetadata(),
-});
