+++ src/data/test/mock-http.ts
@@ -4,6 +4,12 @@ import {
   HttpResponse,
   HttpStatusCode,
 } from '@/data/protocols/http';
+import { faker } from '@faker-js/faker';
+
+export const mockPostRequest = (): HttpPostParams<any> => ({
+  url: faker.internet.url(),
+  body: faker.getMetadata(),
+});
 
 export class HttpPostClientSpy<T, R> implements HttpPostClient<T, R> {
   url?: string;
