+++ src/data/usecases/save-access-token/local-save-access-token.spec.ts
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
 import { LocalSaveAccessToken } from './local-save-access-token';
-import { SetStorageMock } from '@/data/test/mock-storage';
+import { SetStorageMock } from '@/data/test';
 
 type SutTypes = {
   sut: LocalSaveAccessToken;
